+++ pylinac/core/mtf.py
@@ -3,12 +3,16 @@ from __future__ import annotations
 import math
 import warnings
 from collections.abc import Sequence
+from typing import Any, Callable, Literal
 
 import argue
 import numpy as np
 import plotly.graph_objects as go
 from matplotlib import pyplot as plt
+from numpy import ndarray
+from scipy.fft import fft, fftfreq
 from scipy.interpolate import interp1d
+from scipy.signal import windows
 
 from .contrast import michelson
 from .plotly_utils import add_title
@@ -299,3 +303,154 @@ class MomentMTF:
         spacing_ax.set_xlabel("Pair Distance (mm)")
         plt.tight_layout()
         return axis
+
+
+class EdgeSpreadFunctionMTF:
+    """This class will calculate relative MTF from multiple edge spread function (ESF)
+    The MTF is calculated for each ESF and the output is the average of all.
+
+    Parameters
+    ----------
+    esf : list[ndarray]
+        These are the edge spread functions (ESF). Each element of the list represents an ESF.
+    sample_spacing : float | None
+        This is the sample spacing in mm. If None, the frequency axis is cycles/pixel, otherwise it's converted to lp/mm. Default is None.
+    padding_mode : Literal["none", "fixed", "auto"]
+        This is the padding mode (adding zeros) to increase resolution. Default is "auto"
+
+        * mode="none": array is unchanged
+        * mode="fixed": pad to ``num_samples`` (must be larger than the largest array)
+        * mode="auto": pad to the next power of two from the number of samples and ``num_samples``
+    num_samples : int
+        This is the size of the array after padding. Only applicable if padding_mode is "fixed" or "auto". Default is 1024
+    windowing : Callable | None
+        This is the function used to window the ESF. Default is Hann window.
+    kwargs
+        These are the parameters to be used when calling ``windowing``, ie windowing(kwargs)
+    """
+
+    def __init__(
+        self,
+        esf: list[ndarray],
+        sample_spacing: float | None = None,
+        padding_mode: Literal["none", "fixed", "auto"] = "auto",
+        num_samples: int = 1024,
+        windowing: Callable | None = windows.hann,
+        **kwargs: Any,
+    ):
+        self.sample_spacing = sample_spacing
+
+        # boxcar window is just a sequence of '1' so is the same as doing nothing
+        windowing = windowing or windows.boxcar
+
+        len_esf = np.unique([len(e) for e in esf])
+        if padding_mode == "none":
+            # validate that all arrays are the same size
+            if len(len_esf) > 1:
+                raise ValueError(
+                    "If padding_mode='none', all ESF samples must have the same size"
+                )
+            num_samples = len_esf[0]
+        elif padding_mode == "fixed":
+            # validate that num_samples is larger the largest array
+            if num_samples < max(len_esf):
+                raise ValueError("num_samples must be larger than the largest array")
+        elif padding_mode == "auto":
+            # Select the next power of two (or num_samples if larger)
+            next_power_of_two = max(2 ** np.ceil(np.log2(len_esf)))
+            num_samples = int(max(next_power_of_two, num_samples))
+
+        # frequency axis (lp/mm)
+        pixel_spacing = 1 if sample_spacing is None else sample_spacing
+        freq = fftfreq(num_samples, d=pixel_spacing)
+        self.freq = freq[: num_samples // 2]
+
+        # individual results
+        results = [_compute_esf_mtf(e, num_samples, windowing, **kwargs) for e in esf]
+        self._mtf, self._esf, self._lsf, self._lsf_windowed = (
+            list(x) for x in zip(*results)
+        )
+
+        # overall mtf
+        self.mtf = np.mean(np.array(self._mtf), axis=0)
+
+    @argue.bounds(x=(0, 100))
+    def relative_resolution(self, x: float = 50) -> float:
+        """Return the line pair value at the given rMTF resolution value.
+
+        Parameters
+        ----------
+        x : float
+            The percentage of the rMTF to determine the line pair value. Must be between 0 and 100.
+        """
+        # invert x and mtf since interp requires xp to be increasing
+        return float(np.interp(-x / 100, -self.mtf, self.freq))
+
+    def plot(
+        self,
+        axis: plt.Axes | None = None,
+        grid: bool = True,
+        x_label: str | None = None,
+        y_label: str = "Relative MTF",
+        title: str = "RMTF",
+        margins: float = 0.05,
+        label: str = "rMTF",
+    ) -> list[plt.Line2D]:
+        if x_label is None:
+            x_label = (
+                "Cycles / sample" if self.sample_spacing is None else "Line pairs / mm"
+            )
+
+        if axis is None:
+            fig, axis = plt.subplots()
+        points = axis.plot(self.freq, self.mtf, label=label)
+        axis.margins(margins)
+        axis.grid(grid)
+        axis.set_xlabel(x_label)
+        axis.set_ylabel(y_label)
+        axis.set_title(title)
+        plt.tight_layout()
+        return points
+
+    def _plot_debug(self, plot_together: bool = False):
+        n_esf = len(self._mtf)
+        rows = 5
+        cols = 1 if plot_together else n_esf
+
+        fig, axis = plt.subplots(rows, cols)
+        axis = (axis[np.newaxis]).transpose() if cols == 1 else axis
+        for esf_idx in range(n_esf):
+            col_idx = 0 if cols == 1 else esf_idx
+            ax = axis[0, col_idx]
+            ax.set_title("ESF")
+            ax.plot(self._esf[esf_idx])
+
+            ax = axis[1, col_idx]
+            ax.set_title("LSF")
+            ax.plot(self._lsf[esf_idx])
+
+            ax = axis[2, col_idx]
+            ax.set_title("LSF windowed")
+            ax.plot(self._lsf_windowed[esf_idx])
+
+            ax = axis[3, col_idx]
+            ax.set_title("MTF")
+            ax.plot(self.freq, self._mtf[esf_idx])
+
+            ax = axis[4, col_idx]
+            ax.set_title("MTF - diff")
+            ax.plot(self.freq, self._mtf[esf_idx] - self.mtf)
+
+        plt.tight_layout()
+        plt.show()
+
+
+def _compute_esf_mtf(
+    esf: ndarray, num_samples: int, windowing: Callable, **kwargs
+) -> tuple[ndarray, ndarray, ndarray, ndarray]:
+    lsf = np.gradient(esf)
+    lsf_windowed = lsf * windowing(len(esf), **kwargs)
+    mtf = np.abs(fft(lsf_windowed, num_samples))
+    mtf /= mtf[0]  # Normalize
+    mtf = mtf[: num_samples // 2]
+    return mtf, esf, lsf, lsf_windowed